Rename api_service -> apiService (internal casing)
Phase 5: internal-naming consistency (decision #5). Rename the module file
and update all 26 importers + its test file. Mechanical, no behavior change;
the HTTP wire format and all snake_case field names are untouched.
