+++ frontend/src/hooks/useNextUp.ts
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useState } from 'react';
-import { addToNextUp, clearNextUp, fetchNextUp, removeFromNextUp } from '../utils/api_service';
+import { addToNextUp, clearNextUp, fetchNextUp, removeFromNextUp } from '../utils/apiService';
 
 export function useNextUp() {
   const [ids, setIds] = useState<string[]>([]);
